requirements.txt
requests==2.26.0
streamlit==1.3.1
speedtest-cli==2.1.3
pandas==1.3.5
playwright==1.17.2